+++ apps/workers/suppliers/saw.rb
@@ -11,7 +11,7 @@ module Workers::Suppliers
     end
 
     def perform
-      result = importer.fetch_countries
+      result = synchronisation.new_context { importer.fetch_countries }
 
       if result.success?
         countries = result.value
@@ -21,7 +21,9 @@ module Workers::Suppliers
         return
       end
 
-      result = importer.fetch_available_properties_by_countries(countries)
+      result = synchronisation.new_context do
+        importer.fetch_properties_by_countries(countries)
+      end
 
       if result.success?
         properties = result.value
@@ -31,52 +33,62 @@ module Workers::Suppliers
         return
       end
 
-      result = importer.fetch_all_unit_rates_for_properties(properties)
+      result = synchronisation.new_context do
+        importer.fetch_all_unit_rates_for_properties(properties)
+      end
 
       if result.success?
         all_unit_rates = result.value
       else
-        message = "Failed to perform the `#fetch_rates_for_properties` operation"
+        message = "Failed to perform the `#fetch_all_unit_rates_for_properties` operation"
         announce_error(message, result)
         return
       end
 
       properties.each do |property|
-        synchronisation.start(property.internal_id) do
-          Concierge.context.disable!
+        result = synchronisation.new_context do
+          importer.fetch_detailed_property(property.internal_id)
+        end
+        if result.success?
+          detailed_property = result.value
 
-          unit_rates = find_rates(property.internal_id, all_unit_rates)
-          fetch_details_and_build_property(property, unit_rates)
+          next if skip?(detailed_property)
+
+          synchronisation.start(property.internal_id) do
+            unit_rates = find_rates(property.internal_id, all_unit_rates)
+            Result.new ::SAW::Mappers::RoomoramaProperty.build(
+              property,
+              detailed_property,
+              unit_rates
+            )
+          end
+        else
+          synchronisation.failed!
+          # potentially a more meaningful result can be passed from HTTPClient, into result.error.data
+          announce_error("Failed to perform the `#fetch_detailed_property` operation", result)
         end
       end
 
       synchronisation.finish!
     end
 
-    def fetch_details_and_build_property(property, rates)
-      result = importer.fetch_detailed_property(property.internal_id)
-
-      if result.success?
-        detailed_property = result.value
-
-        roomorama_property = ::SAW::Mappers::RoomoramaProperty.build(
-          property,
-          detailed_property,
-          rates
-        )
-
-        Result.new(roomorama_property)
-      else
-        message = "Failed to perform the `#fetch_detailed_property` operation"
-        announce_error(message, result)
-        result
-      end
-    end
-
     private
 
+    # Check if we can skip(and not publish) property, because of the following cases
+    #   - Postal code is "."
+    #
+    # Returns true to caller if skipped
+    #
+    def skip?(detailed_property)
+      if detailed_property.postal_code == "."
+        synchronisation.skip_property(detailed_property.internal_id, "Invalid postal_code: .")
+        return true
+      end
+      return false
+    end
+
     def importer
-      @properties ||= ::SAW::Importer.new(credentials)
+      @importer ||= ::SAW::Importer.new(credentials)
     end
 
     def credentials
@@ -101,7 +113,7 @@ module Workers::Suppliers
     end
 
     def find_rates(property_id, all_unit_rates)
-      all_unit_rates.find { |rate| rate.id == property_id.to_s }
+      all_unit_rates.find { |rate| rate.property_id == property_id.to_s }
     end
   end
 end