+++ spec/workers/suppliers/saw_worker_spec.rb
@@ -51,6 +51,7 @@ RSpec.describe Workers::Suppliers::SAW do
       mock_request(:country, :one)
       mock_request(:propertysearch, :success)
       mock_request(:propertyrates, :success)
+      mock_request(:propertydetail, :success)
 
       expected_property_ids = [1787, 1757, 2721, 2893, 1766]
 
@@ -63,34 +64,47 @@ RSpec.describe Workers::Suppliers::SAW do
       expect(result.to_h[:successful]).to be true
     end
 
-    it "fails when there is an error while fetching property details" do
-      mock_request(:propertydetail, :error)
+    describe "handling property detail" do
+      before do
+        mock_request(:country, :one)
+        mock_request(:propertysearch, :success)
+        mock_request(:propertyrates, :success)
+      end
+      it "fails when there is an error while fetching property details" do
+        mock_request(:propertydetail, :error)
 
-      property = SAW::Entities::BasicProperty.new
-      result = worker.fetch_details_and_build_property(property, nil)
+        result = worker.perform
 
-      expect(result).to be_kind_of(Result)
-      expect(result).not_to be_success
-      expect(result.error.code).to eq("0000")
-      expect(last_context_event[:label]).to eq(
-        "Synchronisation Failure"
-      )
-      expect(last_context_event[:message]).to eq(
-        "Failed to perform the `#fetch_detailed_property` operation"
-      )
-      expect(last_context_event[:backtrace]).to be_kind_of(Array)
-      expect(last_context_event[:backtrace].any?).to be true
-    end
+        expect(result).to be_a(SyncProcess)
+        expect(result.successful).to eq false
+        expect(last_context_event[:label]).to eq(
+          "Synchronisation Failure"
+        )
+        expect(last_context_event[:message]).to eq(
+          "Failed to perform the `#fetch_detailed_property` operation"
+        )
+        expect(last_context_event[:backtrace]).to be_kind_of(Array)
+        expect(last_context_event[:backtrace].any?).to be true
+      end
 
-    it "returns built property" do
-      mock_request(:propertydetail, :success)
+      it "returns built property" do
+        mock_request(:propertydetail, :success)
 
-      property = SAW::Entities::BasicProperty.new
-      result = worker.fetch_details_and_build_property(property, nil)
+        expect_any_instance_of(Workers::PropertySynchronisation).to receive(:process).exactly(5).times
+        result = worker.perform
 
-      expect(result).to be_kind_of(Result)
-      expect(result).to be_success
-      expect(result.value).to be_kind_of(Roomorama::Property)
+        expect(result).to be_a(SyncProcess)
+        expect(result.successful).to eq true
+      end
+
+      it "returns a built property with invalid postal_code" do
+        mock_request(:propertydetail, :invalid_postal_code)
+        expect_any_instance_of(Workers::PropertySynchronisation).to receive(:skip_property).exactly(5).times
+        result = worker.perform
+
+        expect(result).to be_a(SyncProcess)
+        expect(result.successful).to eq true
+      end
     end
   end
 end